Fix OID confusion when excluding extension
When checking for dependencies between an object and an extension we
ensure the referenced object is an extension. However, we were missing a
check that the dependent object was the correct type. An exception was
`functions.sql` which already checked the dependent was a `pg_proc`.

When run against a system catalog that has OID collisions between
database objects this can falsely exclude the objects as internal. This
usually will crash shortly after due to excluding indexes, relations,
etc. that reference each other in the schema.

Fix by extending the existing `functions.sql` check to the other
database object types.