Add exclusion of partition child tables in privileges.sql
The code is modified to exclude the partition child tables when checking for table privileges in the `privileges.sql` file. Earlier, the privileges check was including all the tables. With this change, the partitioned child tables, which generally do not require independent privileges, are excluded from the privileges check. This enhances the accuracy of the privilege validation process.