Add support for inspecting aggregate functions
This commit adds new functionality to the codebase to allow inspection of aggregate database functions. The `InspectedAggFunction` class was added to `schemainspect/pg/obj.py`, with properties needed to create, describe, and delete aggregate functions. Moreover, `load_aggregate_functions` method was added to the `PostgreSQL` class in the same file to facilitate loading of these functions. Associated queries for getting this information are stored in `agg_functions.sql`. This was done to enhance the library's capabilities and cover more aspects of database inspecting.